# docs/openapi/cycleforge-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CycleForge V1 Label Ingestions",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reminders": {
      "get": {
        "description": "The caller's task and checklist reminders in [from, from + days), sorted by remindAt. Clients schedule LOCAL notifications and replace every pending one whose id is no longer in the feed. Task reminders are included only when the session holds work_orders.claim.",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "ISO instant; defaults to now.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 14,
              "default": 7
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The session staffer's reminders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRemindersPayload"
                }
              }
            }
          },
          "400": {
            "description": "Invalid from or days",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authenticated device or session required"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_REQUEST",
                  "INVALID_PDF",
                  "PAYLOAD_TOO_LARGE",
                  "CLIENT_EVENT_PAYLOAD_MISMATCH",
                  "INGESTION_NOT_FOUND",
                  "INGESTION_NOT_ACTIONABLE",
                  "ROW_VERSION_CONFLICT",
                  "INGESTION_APPLY_CONFLICT",
                  "INGESTION_PROCESSING_FAILED"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "StaffReminder": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "source",
          "sourceId",
          "title",
          "body",
          "dueAt",
          "remindAt",
          "urgent",
          "deepLink"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable notification id: `task:<id>` or `checklist:<itemId>:<YYYY-MM-DD>`. Use it as the local notification identifier."
          },
          "source": {
            "type": "string",
            "enum": [
              "task",
              "checklist"
            ]
          },
          "sourceId": {
            "type": "integer",
            "description": "work_assignments.id or daily_check_items.id"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": [
              "string",
              "null"
            ]
          },
          "dueAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the work is due; null for a task with no deadline."
          },
          "remindAt": {
            "type": "string",
            "format": "date-time",
            "description": "When to ring. Always inside the requested window."
          },
          "urgent": {
            "type": "boolean"
          },
          "deepLink": {
            "type": "string",
            "description": "Same-origin /m/... path the notification opens."
          }
        }
      },
      "StaffRemindersPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "generatedAt",
              "staffId",
              "from",
              "to",
              "reminders"
            ],
            "properties": {
              "generatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "staffId": {
                "type": "integer"
              },
              "from": {
                "type": "string",
                "format": "date-time"
              },
              "to": {
                "type": "string",
                "format": "date-time"
              },
              "reminders": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StaffReminder"
                }
              }
            }
          }
        }
      }
    }
  }
}